.infrastructure/requirements.txt
# aws-cdk-lib>=2.109.0 for Python 3.12 Lambda runtime, smstudio.domain to be able to set the
# "StudioWebPortal" user setting to force classic Studio experience
# >=2.140 to try and avoid 'Package @aws-sdk/client-cognito-identity-provider does not exist.' on
# AwsCustomResource (this version includes fix for related failure to upgrade AWS SDK)
# See: https://github.com/aws/aws-cdk/issues/30067
aws-cdk-lib==2.158.0
aws-cdk.aws-lambda-python-alpha==2.158.0-alpha.0
cdk-nag==2.28
constructs>=10.0.0,<11.0.0
upsert-slr>=1.0.2,<2
